pyproject.toml
[tool.poetry]
name = "data-analysis"
version = "0.1.1"
description = "UBC Solar's data analysis environment"
authors = ["Joshua Riefman <joshuariefman@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
jupyterlab = "^4.2.4"
scipy = "^1.14.0"
ipykernel = "^6.29.5"
ubc-solar-physics = "^1.2.0"
ubc-solar-data-tools = "^1.3.1"
numpy = "^2.0.1"
seaborn = "^0.13.2"
scikit-learn = "^1.5.2"
scipy-stubs = "^1.15.2.2"
pvlib = "^0.11.2"
setuptools = "^80.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "matplotlib>=3.10.8",
    "ubc-solar-data-tools>=1.9.1",
]
